refactor(api): cle domain (#4002)
* refactor(api): cle domain

* refactor(api): cle domain import and test refactor

* refactor(api): after main rebase
