+++ api/src/controllers/cle/index.js
@@ -1,8 +1,6 @@
 const express = require("express");
 const router = express.Router();
 
-router.use("/classe", require("./classe"));
-router.use("/etablissement", require("./etablissement"));
 router.use("/referent", require("./referent"));
 router.use("/referent-signup", require("./referent-signup"));
 router.use("/young", require("./young"));
